feat: Add isDev configuration option for database initialization behavior
- Introduced `isDev` boolean in `config.yaml` to control database schema migrations.
- Implemented conditional schema migrations for MySQL based on `isDev` value.
- Enhanced database initialization logic in `internal/model/init.go` to support both SQLite and MySQL.
- Updated `Track`, `TrackPlayRecord`, and `Genre` models to use int64 and int8 types for IDs and track numbers.
- Modified scrobbling logic to ensure compatibility with new data types.
- Improved error logging for failed scrobbles.
- Updated HTML templates for better UI presentation.
- Added documentation for `isDev` feature in the memory index.

+++ cmd/sync_records.go
@@ -56,7 +56,7 @@ func syncRecords(cmd *cobra.Command, args []string) error {
 			Duration:           record.Duration,
 			Timestamp:          record.PlayTime.Unix(),
 			MusicBrainzTrackID: record.MusicBrainzID,
-			TrackNumber:        record.TrackNumber,
+			TrackNumber:        int64(record.TrackNumber),
 		}
 
 		_, err := lastfm.PushTrackScrobble(context.Background(), req)
